[omnibus] Ensure datadog-agent-prepare is processed first
Omnibus ensures that the first dependency component in the project file
is always processed first.
All other components in the project file are processed in the normal
order, which is: after all dependencies of all components of the project
file are processed.

Since datadog-agent-prepare is not the first component in the project
on linux, and datadog-agent is a dependency of datadog-agent-integrations-py*,
we end up with datadog-agent being processed before datadog-agent-prepare
on linux, making the latter pointless.

However, the build does work with this order, so is the datadog-agent-prepare
definition of any use on linux (and maybe on other distros)?