pkg/trace/obfuscate: handle escape characters in SQL based on engine (#4327)
This commit fixes an issue, where our current SQL tokenizer assumes that a backslash within a string always acts as an escape character. In dialects such as MySQL, this turns out to be a good assumption, but in general, backslashes in SQL strings are treated literally.

The following query should be valid, but if we treat the backslash as an escape character, the tokenizer will fail to parse it.

    SELECT * FROM foo WHERE name = 'backslash\' AND id ='1234';

This commit changes the behavior of the tokenizer to first assume that backlashes escape quotes, but if we eventually fail to parse the entire query, we will retry tokenization assuming that backslashes are treated literally. The big assumption here is that within a single query, backslashes will be treated consistently.