[logs/docker] configurable docker client read timeout.
The Docker daemon could have a high latency replying to a docker logs
client call while it's gathering timestamps on several large log files.
Configuring this field to a higher duration could help solve this situation.
Note that a better solution is to configure the docker daemon to not store
too much large log files.